+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Added google mobile ads sdk
+* Added easy ads singleton class to show interstitial and rewarded video ad
+
+## 0.0.2
+
+* Added Changelog
+* Added how to use